+++ lib/service/LiveStats.dart
@@ -20,7 +20,7 @@ import 'package:rokwire_plugin/service/app_livecycle.dart';
 import 'package:illinois/model/livestats/LiveGame.dart';
 import 'package:rokwire_plugin/service/log.dart';
 import 'package:illinois/service/Config.dart';
-import 'package:illinois/service/Network.dart';
+import 'package:rokwire_plugin/service/network.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/service.dart';
 import 'package:illinois/service/Storage.dart';
@@ -164,7 +164,7 @@ class LiveStats with Service implements NotificationsListener {
 
   void _loadLiveGames() {
     String? url = (Config().sportsServiceUrl != null) ? "${Config().sportsServiceUrl}/api/v2/live-games" : null;
-    var response = Network().get(url, auth: NetworkAuth.Auth2);
+    var response = Network().get(url, auth: NetworkAuth.auth2);
     response.then((response) {
     String? responseBody = response?.body;
       if ((response != null) && (response.statusCode == 200)) {